use include_bytes! for less memory overhead
This uses `include_bytes!` to generate the bytestring rather than
the `quote!` macro.  With this change, the total memory used by rustc
for a crate that includes a 15 MB file drops from over 8 GB to only
about 700 MB.

This is similar to #23, but that predated the proc_macro.

It should also resolve *some* aspects of #31 (namely for modified
files), but not all. (the crate still has no way to track new files)