Rewrite find files in O(lg(n)), and clean up the api
This PR makes some large changes to the library, in order to
reduce finding a file from O(n) to O(lg(n)), by maintaining
the directory entries in a sorted order, and binary searching
for a file.

In order to accomplish this, I changed how directories
work. Instead of each directory supporting the ability to look
up a path relative to the root directory, I instead introduce
a `FileSystem` that is the entrypoint for resolving a path into
contents. This is more analogous to how std::fs operates.