Relax returned lifetimes of get_dir and get_file
Prior to this change, the elided lifetime '_ would be constrained to
at most the lifetime of the &self reference _or_ the internal lifetime
'a, when instead it could just be 'a.

If &self isn't actually behind a static, but 'a: 'static, the old behavior
leads to some rather puzzling lifetime errors.